airflow_folder/requirements.txt
pandas==1.3.5
cassandra-driver==3.25.0